Use `$(MAKE)` instead of hard-coding the command name of GNU make
[Why]
GNU make be installed as `gmake` for instance, to avoid a conflict with
other make(1) implementations.

This is the case on FreeBSD and other BSDs for instance.

[How]
All make(1) implementations expose the name of the actual make(1)
executable in the `$(MAKE)` variable. This way, we are sure to honor
whatever command the user executed.